Refactor user model and enhance notifications: replace notifications structure with INotification interface, improve sidebar layout, and add notifications component to dashboard
