+++ package.json
@@ -27,7 +27,7 @@
     "axios": "^1.7.9",
     "bcryptjs": "^2.4.3",
     "chart.js": "^4.4.7",
-    "framer-motion": "^11.14.0",
+    "framer-motion": "^11.14.1",
     "jose": "^5.9.6",
     "lodash": "^4.17.21",
     "mongoose": "^8.8.4",
@@ -53,7 +53,7 @@
     "@typescript-eslint/eslint-plugin": "^8.18.0",
     "@typescript-eslint/parser": "^8.18.0",
     "autoprefixer": "^10.4.20",
-    "daisyui": "^4.12.20",
+    "daisyui": "^4.12.22",
     "eslint": "^9.16.0",
     "eslint-plugin-react": "^7.37.2",
     "globals": "^15.13.0",
